Agents added (#84)
* refactoring

* refactoring

* refactoring: wip

* refactoring: wip

* refacotring

* refactoring

* llms added
