+++ app/src/client/admin/pages/Settings.tsx
@@ -18,7 +18,7 @@ const Settings = () => {
   return (
     <DefaultLayout>
       <div className="mx-auto max-w-270">
-        
+
         <Breadcrumb pageName="Settings" />
 
         <div className="grid grid-cols-5 gap-8">
