+++ app/src/client/hooks/useColorMode.tsx
@@ -15,4 +15,3 @@ export default function useColorMode() {
 
   return [colorMode, setColorMode];
 };
-
