+++ docs/docs/en/api/fastagency/models/base/ObjectReference.md
@@ -8,4 +8,4 @@ search:
   boost: 0.5
 ---
 
-::: fastagency.models.registry.get_wrapper_model
+::: fastagency.models.base.ObjectReference
