+++ docs/docs/en/api/fastagency/models/base/ObjectWrapper.md
@@ -8,4 +8,4 @@ search:
   boost: 0.5
 ---
 
-::: fastagency.models.registry.ObjectWrapper
+::: fastagency.models.base.ObjectWrapper
