+++ docs/docs/en/api/fastagency/models/base/create_reference_model.md
@@ -8,4 +8,4 @@ search:
   boost: 0.5
 ---
 
-::: fastagency.models.registry.ObjectReference
+::: fastagency.models.base.create_reference_model
