+++ docs/docs/en/api/fastagency/models/llms/azure/AzureOAIAPIKey.md
@@ -8,4 +8,4 @@ search:
   boost: 0.5
 ---
 
-::: fastagency.models.llms.AzureOAI
+::: fastagency.models.llms.azure.AzureOAIAPIKey
