+++ fastagency/models/agents/__init__.py
@@ -1,4 +1,1 @@
-from ._assistant import AssistantAgent
-from ._web_surfer import WebSurferAgent
-
-__all__ = ["AssistantAgent", "WebSurferAgent"]
+from . import assistant, web_surfer  # noqa: F401
