+++ fastagency/models/llms/openai.py
@@ -1,23 +1,20 @@
 from typing import Annotated, Literal
 
-from pydantic import BaseModel, Field, HttpUrl, model_validator
+from pydantic import Field, HttpUrl, model_validator
 from typing_extensions import Self, TypeAlias
 
 from ...constants import OPENAI_MODELS_LITERAL
-from ..registry import get_reference_model, get_wrapper_model, register
+from ..base import Model
+from ..registry import register
 
 __all__ = [
     "OpenAIAPIKey",
-    "OpenAIAPIKeyRef",
-    "OpenAIAPIKeyWrapper",
     "OpenAI",
-    "OpenAIRef",
-    "OpenAIWrapper",
 ]
 
 
 @register("secret")
-class OpenAIAPIKey(BaseModel):
+class OpenAIAPIKey(Model):
     api_key: Annotated[str, Field(description="The API Key from OpenAI")]
 
     @model_validator(mode="after")
@@ -28,12 +25,11 @@ class OpenAIAPIKey(BaseModel):
         return self
 
 
-OpenAIAPIKeyRef: TypeAlias = get_reference_model(OpenAIAPIKey)  # type: ignore[valid-type]
-OpenAIAPIKeyWrapper: TypeAlias = get_wrapper_model(OpenAIAPIKey)  # type: ignore[valid-type]
+OpenAIAPIKeyRef: TypeAlias = OpenAIAPIKey.get_reference_model()  # type: ignore[valid-type]
 
 
 @register("llm")
-class OpenAI(BaseModel):
+class OpenAI(Model):
     model: Annotated[  # type: ignore[valid-type]
         Literal[OPENAI_MODELS_LITERAL],
         Field(description="The model to use for the OpenAI API, e.g. 'gpt-3.5-turbo'"),
@@ -49,7 +45,3 @@ class OpenAI(BaseModel):
         Literal["openai"],
         Field(title="API Type", description="The type of the API, must be 'openai'"),
     ] = "openai"
-
-
-OpenAIRef: TypeAlias = get_reference_model(OpenAI)  # type: ignore[valid-type]
-OpenAIWrapper: TypeAlias = get_wrapper_model(OpenAI)  # type: ignore[valid-type]
